+++ assistant/src/__tests__/gmail-integration.test.ts
@@ -29,8 +29,7 @@ describe("Messaging tool contract", () => {
     "gmail_trash",
     "gmail_draft",
     "gmail_unsubscribe",
-    "gmail_list_attachments",
-    "gmail_download_attachment",
+    "gmail_attachments",
     "gmail_send_with_attachments",
     "gmail_forward",
     "gmail_follow_up",
